Remove problematic APIs
Needing to store every handle in a Map is problematic for several
reason. So the API has been removed and replaced by passing the handle
as the fourth argument to init(). This will still allow users to
implement a similar feature on their own, without placing a performance
burden directly on node core.